requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
llama-cpp-python==0.2.20
pydantic==2.5.0
aiohttp==3.9.1

